backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.114.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
sqlalchemy = "^2.0.34"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.30.6"
ipython = "^8.27.0"
alembic = "^1.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
